Elevate all parts of Play.js into App.js
All game state and associated clickHandlers now resides at the App level.
